+++ src/components/steps/OnSiteCell.tsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { View, StyleSheet } from 'react-native';
 import { CourseSlotType } from 'types/CourseSlotType';
 import { StepType } from 'types/StepType';
-import moment from '../../../core/helpers/moment';
+import moment from '../../core/helpers/moment';
 import CalendarIcon from '../CalendarIcon';
-import { PADDING, BORDER_WIDTH } from '../../../styles/metrics';
-import { GREY } from '../../../styles/colors';
+import { PADDING, BORDER_WIDTH } from '../../styles/metrics';
+import { GREY } from '../../styles/colors';
 import StepCellTitle from './StepCellTitle';
 
 interface OnSiteCellProps {
